Fixed duplicated OSC output #1 & #2 in DP 10 on MacOS (issue #144).
This is a half-baked fix for a really concerning bug...

When DP loads a projet that uses ControlGRIS, it will sometimes load the plugin once more than it should. This fantom instance is not attached to the audio graph, it is never destroyed and it only gets cleaned up by the OS upon closing DP.

This bug seems to have been lying around way before 1.1.6, but it was only picked up then since the OSC sender's initialization was moved into the AudioProcessor CTOR, making the fantom instance send continuous OSC messages to SpatGRIS.

Assigning the plugin to a track in a new project does not trigger the bug, nor does loading this newly-saved projet. It is not clear why DP does this and it's annoying the heck out of me.

I suspect that the problem is lying on Digital Performer's side. To be continued.